Updated a bunch of stuff to intall my config onto any new machine. Installes emacs prelude and the appropriate settings, liquidprompt (though there is some configuration that is not right) and vimrc and oh-my-zsh.

+++ setup.sh
@@ -1,28 +1,51 @@
 #!/usr/bin/env bash
-
-if [ -f $HOME/.bashrc ] ; then
-    printf "Current .bashrc will be removed!"
-    printf "Press ENTER to continue. Ctrl-C to quit"
-    read ENTER
-fi
-rm $HOME/.bashrc
+[ -f $HOME/.bashrc ] && rm $HOME/.bashrc
 ln -s $PWD/.bashrc $HOME/.bashrc
+[ -f $HOME/.inputrc ] && rm $HOME/.inputrc 
+ln -s $PWD/.inputrc $HOME/.inputrc
+[ -f $HOME/.bash_logout ] && rm $HOME/.bash_logout 
+ln -s $PWD/.bash_logout $HOME/.bash_logout
+[ -f $HOME/.profile ] && rm $HOME/.profile
+ln -s $PWD/.profile $HOME/.profile
 
-if [[ -d $HOME/liquidprompt ]] ; then
-    printf "Looks like you have liquid prompt installed."
-    printf "Symlinking .liquidpromptrc"
-    rm $HOME/.liquidpromptrc
-    ln -s $PWD/.liquidpromptrc $HOME/.liquidpromptrc
+
+#Install liquidprompt and set it up
+if [[ ! -d $HOME/liquidprompt ]] ; then
+    printf "Cloning liquidprompt from github\n"
+    git clone https://github.com/nojhan/liquidprompt.git $HOME/liquidprompt
 fi
+printf "Symlinking .liquidpromptrc\n"
+[ -f $HOME/.liquidpromptrc ] && rm $HOME/.liquidpromptrc
+ln -s $PWD/.liquidpromptrc $HOME/.liquidpromptrc
 
-if [[ -f $HOME/.bash_logout ]] ; then
-    rm $HOME/.bash_logout
-    printf "Symlinking new .bash_logout"
-    ln -s $PWD/.bash_logout $HOME/.bash_logout
-fi
-
-
-if [[ $SHELL == '/bin/zsh' ]]; then
-    printf "Horray ZSH. Linking .zshrc"
+# check if zsh is installed
+which zsh &>/dev/null  && ZSH_INSTALLED=1
+if [ $ZSH_INSTALLED ] ; then 
+    if [[ ! -d $HOME/.oh-my-zsh ]] ; then
+        printf "Cloning oh-my-zsh from github\n"
+        git clone https://github.com/robbyrussell/oh-my-zsh.git $HOME/.oh-my-zsh
+    fi
+    printf "Horray ZSH. Linking .zshrc\n"
+    [ -f $HOME/.zshrc ] && rm $HOME/.zshrc 
     ln -s $PWD/.zshrc       $HOME/.zshrc
 fi
+
+#Install emacs prelude 
+if [[ ! -d $HOME/prelude ]] ; then 
+    printf "Cloning emacs prelude from github\n"
+    git clone https://github.com/bbatsov/prelude.git $HOME/prelude
+    ln -s $HOME/prelude $HOME/.emacs.d
+fi
+# symlink the emacs personal files
+[ -d $HOME/prelude/personal/emacs ] && rm -rf $HOME/prelude/personal/emacs
+[ -f $HOME/prelude/prelude-modules.el ] && rm $HOME/prelude/prelude-modules.el
+ln -s $PWD/prelude-modules.el $HOME/prelude/prelude-modules.el
+
+for file in $PWD/emacs/* ; do
+    ln -s $PWD/emacs/$(basename $file) $HOME/prelude/personal/$(basename $file) ;
+done
+
+# symlink the tmux config
+# symlink .vimrc
+[ -f $HOME/.vimrc ] && rm $HOME/.vimrc
+ln -s $PWD/.vimrc $HOME/.vimrc
